api/requirements.txt
flask
flask_sqlalchemy
werkzeug
nanoid
Flask-Migrate
psycopg2-binary
python-dotenv
sqlalchemy_utils
pyjwt
waitress
flask-cors
pyopenssl
flask-marshmallow
marshmallow-sqlalchemy
requests
flask[async]
aiohttp
flask-socketio
simple-websocket